requirements.txt
fastapi[all]
sqlalchemy
psycopg2
pydantic_settings
passlib[bcrypt]
python-jose[cryptography]